# core-lending-supply-chain-finance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-supply-chain-finance API
  description: Supply Chain Finance Management System (Factoring & Confirming)
  contact:
    name: Firefly Software Foundation
    url: https://getfirefly.io
    email: dev@getfirefly.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
...
paths:
  /api/v1/supply-chain-finance/settlements/{id}:
    get:
      tags:
        - SupplyChainFinanceSettlement
      summary: Get a supply chain finance settlement by ID
      description: Retrieve a specific settlement by its unique identifier
      operationId: getById
      parameters:
        - name: id
          in: path
          description: Unique identifier of the settlement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Settlement found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        '404':
          description: Settlement not found
    put:
      tags:
        - SupplyChainFinanceSettlement
      summary: Update an existing supply chain finance settlement
      description: Update the details of an existing settlement
      operationId: update
      parameters:
        - name: id
          in: path
          description: Unique identifier of the settlement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        required: true
      responses:
        '200':
          description: Settlement updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        '400':
          description: Invalid settlement data
        '404':
          description: Settlement not found
    delete:
      tags:
        - SupplyChainFinanceSettlement
      summary: Delete a supply chain finance settlement
      description: Delete a settlement from the system
      operationId: delete
      parameters:
        - name: id
          in: path
          description: Unique identifier of the settlement
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Settlement deleted successfully
        '404':
          description: Settlement not found
  /api/v1/supply-chain-finance/invoices/{id}:
    get:
      tags:
        - SupplyChainFinanceInvoice
      summary: Get a supply chain finance invoice by ID
      description: Retrieve a specific invoice by its unique identifier
      operationId: getById_1
      parameters:
        - name: id
          in: path
          description: Unique identifier of the invoice
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Invoice found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        '404':
          description: Invoice not found
    put:
      tags:
        - SupplyChainFinanceInvoice
      summary: Update an existing supply chain finance invoice
      description: Update the details of an existing invoice
      operationId: update_1
      parameters:
        - name: id
          in: path
          description: Unique identifier of the invoice
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        required: true
      responses:
        '200':
          description: Invoice updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        '400':
          description: Invalid invoice data
        '404':
          description: Invoice not found
    delete:
      tags:
        - SupplyChainFinanceInvoice
      summary: Delete a supply chain finance invoice
      description: Delete an invoice from the system
      operationId: delete_1
      parameters:
        - name: id
          in: path
          description: Unique identifier of the invoice
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Invoice deleted successfully
        '404':
          description: Invoice not found
...
  /api/v1/supply-chain-finance/settlements:
    get:
      tags:
        - SupplyChainFinanceSettlement
      summary: List/Search supply chain finance settlements
      description: Retrieve a paginated list of settlements based on filter criteria
      operationId: findAll
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for settlements
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSupplyChainFinanceSettlementDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved settlements
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - SupplyChainFinanceSettlement
      summary: Create a new supply chain finance settlement
      description: Create a new settlement in the system
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        required: true
      responses:
        '200':
          description: Settlement created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        '400':
          description: Invalid settlement data
  /api/v1/supply-chain-finance/invoices:
    get:
      tags:
        - SupplyChainFinanceInvoice
      summary: List/Search supply chain finance invoices
      description: Retrieve a paginated list of invoices based on filter criteria
      operationId: findAll_1
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for invoices
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSupplyChainFinanceInvoiceDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved invoices
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - SupplyChainFinanceInvoice
      summary: Create a new supply chain finance invoice
      description: Create a new invoice in the system
      operationId: create_1
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        required: true
      responses:
        '200':
          description: Invoice created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        '400':
          description: Invalid invoice data
...
components:
  schemas:
    SupplyChainFinanceSettlementDTO:
      required:
        - currency
        - financeType
        - invoiceId
        - settlementAmount
        - settlementDate
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        invoiceId:
          type: string
          format: uuid
        settlementDate:
          type: string
          format: date
        currency:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
            - JPY
            - CNY
        settlementAmount:
          type: number
        advancedAmount:
          type: number
        interestAmount:
          type: number
        feeAmount:
          type: number
        balanceDue:
          type: number
        paymentReference:
          maxLength: 100
          minLength: 0
          type: string
        notes:
          maxLength: 2000
          minLength: 0
          type: string
    SupplyChainFinanceInvoiceDTO:
      required:
        - agreementId
        - counterpartyId
        - currency
        - dueDate
        - financeType
        - invoiceAmount
        - invoiceDate
        - invoiceNumber
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        agreementId:
          type: string
          format: uuid
        counterpartyId:
          type: string
          format: uuid
        invoiceNumber:
          maxLength: 100
          minLength: 0
          type: string
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        currency:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
            - JPY
            - CNY
        invoiceAmount:
          type: number
        status:
          type: string
          enum:
            - REGISTERED
            - PENDING_VERIFICATION
            - VERIFIED
            - APPROVED
            - REJECTED
            - FINANCED
            - CONFIRMED
            - PAID
            - SETTLED
            - CANCELLED
        financedAmount:
          type: number
        financedDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
        verificationNotes:
          maxLength: 2000
          minLength: 0
          type: string
        rejectionReason:
          maxLength: 500
          minLength: 0
          type: string
        purchaseOrderNumber:
          maxLength: 100
          minLength: 0
          type: string
        deliveryNoteNumber:
          maxLength: 100
          minLength: 0
          type: string
        notes:
          maxLength: 2000
          minLength: 0
          type: string
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestSupplyChainFinanceSettlementDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SupplyChainFinanceSettlementDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    FilterRequestSupplyChainFinanceInvoiceDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SupplyChainFinanceInvoiceDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...